requirement.txt
requests
protobuf==3.20
mlagents_envs
swig
gymnasium
gymnasium[atari,accept-rom-license,box2d,mujoco]
mujoco
box2d
box2d-py
box2d-kengz
jax>=0.4.7
einops
dm-haiku
optax
numpy
ray[default]
colabgymrender
cpprb @ git+https://github.com/tinker495/cpprb.git
tensorboardX
imageio-ffmpeg
opencv-python
pre-commit
